Created Beam subdirectory
