+++ PadmeMC/Beam/GNUmakefile
@@ -1,0 +1,19 @@
+# --------------------------------------------------------------
+# GNUmakefile for Beam library
+# --------------------------------------------------------------
+# --------------------------------------------------------------
+# History:
+#
+# Created by Emanuele Leonardi (emanuele.leonardi@roma1.infn.it) 2016-02-11
+#
+# --------------------------------------------------------------
+
+name := Beam
+G4TARGET := PADMEMC
+G4EXLIB := true
+
+ifndef PADMEMC
+  PADMEMC = ../
+endif
+
+include $(PADMEMC)/config/SubDetector.gmk
