+++ PadmeMC/include/DetectorConstruction.hh
@@ -59,6 +59,10 @@ public:
   void EnableStructure(G4String);
   void DisableStructure(G4String);
 
+  // Return some global information used by beam generator
+  G4double GetTargetFrontFaceZ();
+  G4double GetTargetThickness();
+
 private:
 
   void DefineMaterials();
